requirements.txt
torch>=2.2.0
gymnasium>=0.29.1
stable-baselines3>=2.2.1
networkx>=3.2.1
matplotlib>=3.8.2
numpy>=1.26.3
pandas>=2.2.0
seaborn>=0.13.1
hydra-core>=1.3.2
python-louvain>=0.16
pingouin>=0.5.4
neat-python>=0.92 